requirements.txt
matplotlib~=3.2.2
numpy~=1.18.5
seaborn~=0.10.1
sklearn~=0.0
scikit-learn~=0.23.1
scipy~=1.4.1
tensorflow-gpu~=2.3.0
tensorflow_probability